refactor: convert ChatbotVerificationService to service class with implementation and update imports
