+++ decision/src/test/java/pt/estga/decision/listeners/SubmissionEventListenerTest.java
@@ -11,7 +11,7 @@ import pt.estga.content.services.MarkOccurrenceService;
 import pt.estga.decision.services.DecisionServiceFactory;
 import pt.estga.file.entities.MediaFile;
 import pt.estga.monument.Monument;
-import pt.estga.monument.MonumentService;
+import pt.estga.monument.services.MonumentService;
 import pt.estga.submission.entities.MarkOccurrenceSubmission;
 import pt.estga.submission.events.SubmissionAcceptedEvent;
 import pt.estga.submission.repositories.MarkOccurrenceSubmissionRepository;
