+++ monument/src/main/java/pt/estga/monument/controllers/MonumentAdminController.java
@@ -13,8 +13,8 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import pt.estga.monument.Monument;
 import pt.estga.monument.MonumentMapper;
-import pt.estga.monument.MonumentQueryService;
-import pt.estga.monument.MonumentService;
+import pt.estga.monument.services.MonumentQueryService;
+import pt.estga.monument.services.MonumentService;
 import pt.estga.monument.dots.MonumentDto;
 import pt.estga.monument.dots.MonumentRequestDto;
 import pt.estga.sharedweb.exceptions.ResourceNotFoundException;
