+++ monument/src/main/java/pt/estga/monument/services/MonumentService.java
@@ -1,8 +1,10 @@
-package pt.estga.monument;
+package pt.estga.monument.services;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import pt.estga.monument.Monument;
+import pt.estga.monument.MonumentRepository;
 
 import java.util.Optional;
